refactor: address all code smells
